Manage canvas element in Vue
